requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
urllib3>=2.0.0